# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Currency": {
				"properties": {
					"id": {
						"type": "string"
					},
					"isoCode2": {
						"type": "string"
					},
					"isoCode3": {
						"type": "string"
					},
					"isoNumericCode": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"isoCode2",
					"isoCode3",
					"isoNumericCode",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CurrenciesErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"details": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CurrencyRequest": {
				"properties": {
					"isoCode2": {
						"type": "string"
					},
					"isoCode3": {
						"type": "string"
					},
					"isoNumericCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"isoCode2",
					"isoCode3",
					"isoNumericCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "goldsphere-server",
		"version": "1.0.0",
		"description": "The server including the business logic and integration for the goldsphere project",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/currencies": {
			"get": {
				"operationId": "ListCurrencies",
				"responses": {
					"200": {
						"description": "Currencies list",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Currency"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrenciesErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCurrency",
				"responses": {
					"201": {
						"description": "Currency created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Currency"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrenciesErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CurrencyRequest"
							}
						}
					}
				}
			}
		},
		"/currencies/{id}": {
			"put": {
				"operationId": "UpdateCurrency",
				"responses": {
					"200": {
						"description": "Currency updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Currency"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrenciesErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CurrencyRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCurrency",
				"responses": {
					"204": {
						"description": "Currency deleted"
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrenciesErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetCurrencyById",
				"responses": {
					"200": {
						"description": "Currency details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Currency"
								}
							}
						}
					},
					"404": {
						"description": "Currency not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrenciesErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrenciesErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	],
...
}